Extract msgpack adaptor for nlohman::json
